+++ idea-to-gold/.wrangler/tmp/bundle-30tapT/middleware-insertion-facade.js
@@ -1,8 +1,8 @@
-				import worker, * as OTHER_EXPORTS from "D:\\ProjectAI\\Idea_to_gold\\idea-to-gold\\.wrangler\\tmp\\pages-UQtBmd\\functionsWorker-0.6665785653236438.mjs";
+				import worker, * as OTHER_EXPORTS from "D:\\ProjectAI\\Idea_to_gold\\idea-to-gold\\.wrangler\\tmp\\pages-NNRFtg\\functionsWorker-0.8726995034401128.mjs";
 				import * as __MIDDLEWARE_0__ from "C:\\Users\\yilai\\AppData\\Roaming\\npm\\node_modules\\wrangler\\templates\\middleware\\middleware-ensure-req-body-drained.ts";
 import * as __MIDDLEWARE_1__ from "C:\\Users\\yilai\\AppData\\Roaming\\npm\\node_modules\\wrangler\\templates\\middleware\\middleware-miniflare3-json-error.ts";
 
-				export * from "D:\\ProjectAI\\Idea_to_gold\\idea-to-gold\\.wrangler\\tmp\\pages-UQtBmd\\functionsWorker-0.6665785653236438.mjs";
+				export * from "D:\\ProjectAI\\Idea_to_gold\\idea-to-gold\\.wrangler\\tmp\\pages-NNRFtg\\functionsWorker-0.8726995034401128.mjs";
 				const MIDDLEWARE_TEST_INJECT = "__INJECT_FOR_TESTING_WRANGLER_MIDDLEWARE__";
 				export const __INTERNAL_WRANGLER_MIDDLEWARE__ = [
