+++ idea-to-gold/.wrangler/tmp/bundle-7wV3Yq/middleware-insertion-facade.js
